requirements.txt
alembic==1.7.5
aniso8601==9.0.1
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.2
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.27
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
